+++ src/dune/ordered_set_lang.mli
@@ -75,8 +75,6 @@ module Unexpanded : sig
 
   val encode : t -> Dune_lang.t list
 
-  val encode_and_upgrade : t -> Dune_lang.t list
-
   val standard : t
 
   val of_strings : pos:string * int * int * int -> string list -> t