+++ src/jbuild_support/atom.mli
@@ -1,0 +1,1 @@
+val is_valid : string -> bool